fix:move sql_data_type file position
